requirements.txt
astroid==2.4.2
autopep8==1.5.3
beautifulsoup4==4.9.1
bs4==0.0.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-core==1.3.0
google-cloud-firestore==1.7.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pylint==2.5.3
pytz==2020.1
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
